Loading an RVM environment before raking

+++ cookbooks/shapado/recipes/default.rb
@@ -60,6 +60,22 @@ end
 bash "Install gems & bootstrap shapado" do
   cwd ::File.join(node[:nginx][:content_dir], "shapado")
   code <<-EOF
+rvm_bin="#{node[:rvm][:install_path]}/bin/rvm"
+echo "Testing for RVM using $rvm_bin"
+if [ ! -f $rvm_bin ]
+then
+  echo "No RVM installation found, not loading RVM environment"
+  exit 0
+fi
+
+if [[ -s "#{node[:rvm][:install_path]}/environments/default" ]] ; then
+  echo "Found a default RVM environment, loading it now"
+  source "#{node[:rvm][:install_path]}/environments/default"
+else
+  echo "No default RVM environment found, can not continue.  Try setting one with rvm --default"
+  exit 0
+fi
+
 cp config/database.yml.sample config/database.yml
 RAILS_GEM_VERSION=#{node[:rails][:version]} rake asset:packager:build_all
 script/update_geoip
